Fix decompilation bug preventing display list submission
ResolutionCheck() cast taskTicks (s8, initialized to -1) to u8 before
comparing < 2. This turned -1 into 255, making the check always fail
and preventing the game from ever building a display list or submitting
a graphics task. Remove the erroneous (u8) cast so the signed comparison
works as intended (-1 < 2 = true).

Also add debug logging to scheduler task dispatch for diagnostics.

+++ src_pseudo/gfx.cpp
@@ -284,7 +284,7 @@ u8 Graphics::ResolutionCheck(void){
   
   if (((gGfxManager.Hres[0] == gGfxManager.Hres[1]) && (gGfxManager.Vres[0] == gGfxManager.Vres[1]))
      && (gGfxManager.colordepth[0] == gGfxManager.colordepth[1])) {
-    bVar1 = (u8)gGfxManager.taskTicks < 2;
+    bVar1 = gGfxManager.taskTicks < 2;
   }
   else {
     bVar1 = false;
